% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_reef_change.R
\name{assess_reef_change}
\alias{assess_reef_change}
\title{Assess reef change between two NDCI rasters}
\usage{
assess_reef_change(ndci_t1, ndci_t2)
}
\arguments{
\item{ndci_t1}{SpatRaster. NDCI at time 1 (e.g. Summer 2024).}

\item{ndci_t2}{SpatRaster. NDCI at time 2 (e.g. Summer 2025).}
}
\value{
A SpatRaster with pixel-wise NDCI difference (t2 minus t1).
Positive values indicate increased chlorophyll (degradation),
negative values indicate decreased chlorophyll (improvement).
}
\description{
Assess reef change between two NDCI rasters
}
